+++ example/lib/star_shape.dart
@@ -2,7 +2,7 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:flutter_androssy_kits/painters.dart';
+import 'package:flutter_androssy_kits/flutter_androssy_kits.dart';
 
 class StarShapePage extends StatefulWidget {
   const StarShapePage({super.key});