+++ lib/flutter_androssy_kits.dart
@@ -1,4 +1,2 @@
-library flutter_androssy_kits;
-
-export 'painters.dart';
+export 'core/painters.dart';
 export 'widgets.dart';