+++ lib/painters.dart
@@ -1,5 +1,0 @@
-library painters;
-
-export 'painters/dock_painter.dart';
-export 'painters/painter.dart';
-export 'painters/star_painter.dart';