+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.4
+
+* Add androssy progress gesture to make progressing pressed button
+* Remove all external libraries and using androssy instance
+* Replace all callback as native flutter style:
+    - onClick to onTap
+    - onDoubleClick to onDoubleTap
+    - onLongClick to onLongPress
+
 ## 1.0.3
 
 * Add androssy fade and coordinator